movie_back/requirements.txt
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.7
django-cors-headers==3.2.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
entrypoints==0.3
gunicorn==20.0.4
idna==2.8
jedi==0.14.1
Jinja2==2.10.1
jsonschema==3.0.1
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psycopg2==2.8.4
Pygments==2.4.2
PyJWT==1.7.1
pyrsistent==0.15.3
python-dateutil==2.8.0
python-decouple==3.3
pytz==2019.3
pywinpty==0.5.5
pyzmq==18.0.2
qtconsole==4.5.1
requests==2.22.0
Send2Trash==1.5.0
six==1.13.0
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==4.1.4
widgetsnbextension==3.5.0
